refactor(utils): 重构工具类并优化音乐播放功能
-重构文件管理类，改名为 file_tools
- 新增 player_tools 类，用于音乐播放相关功能
- 新增 tipbar_tools 类，用于提示栏相关功能
- 重构 config 类，添加 MAIN_PATH常量
- 更新音乐下载和搜索相关的导入路径
