+++ crawlerCore/searchCore.py
@@ -1,5 +1,6 @@
 import random
 import time
+import config
 
 import requests
 from bs4 import BeautifulSoup
@@ -195,3 +196,11 @@ def search_song_online(search_content, page):
 
     result_list.sync_json()
     return result_list
+
+def searchOnBili(search_content):
+    # 将搜索结果写入json
+    result_info = search_song_online(search_content, config.search_page)
+    temp_list = SongList()
+    temp_list.append_list(result_info)
+    temp_list.unique_by_bv()
+    temp_list.save_list(r"data\search_data.json")
